server/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
whitenoise==6.9.0
